requirements.txt
Flask==3.0.0
Flask-Login==0.6.3
Flask-WTF==1.2.1
Werkzeug==3.0.1
python-dotenv==1.0.0
firebase-admin==6.2.0
Pyrebase4==4.7.1
markdown==3.5.1
python-dateutil==2.8.2
WTForms==3.1.1
gunicorn==21.2.0
email-validator==2.1.0.post1
